SW-1061 Move mail delivery to a separate class
In preparation for adding tests to verify the contents of email notifications,
refactor `EmailService` to extract the functions that interact with actual
email services into a separate class `EmailSender` that we can stub out in the
test code.

No behavior change.
